+++ bundles/net.tourbook/src/net/tourbook/ui/tourChart/TourChart.java
@@ -2339,7 +2339,7 @@ public class TourChart extends Chart implements ITourProvider, ITourMarkerUpdate
                final long pausedTime_Start = allTourPauses.get(currentTourPauseIndex).get(0);
                final long pausedTime_End = allTourPauses.get(currentTourPauseIndex).get(1);
 
-               final long pauseDuration = Math.round((float) (pausedTime_End - pausedTime_Start) / 1000);
+               final long pauseDuration = Math.round((pausedTime_End - pausedTime_Start) / 1000f);
                pauseDurationText = UI.format_hh_mm_ss(pauseDuration);
 
                long previousTourElapsedTime = 0;
@@ -2386,7 +2386,7 @@ public class TourChart extends Chart implements ITourProvider, ITourMarkerUpdate
          final int[] timeSerie = _tourData.timeSerie;
          for (int index = 0; index < pausedTime_Start.length; ++index) {
 
-            final long pauseDuration = Math.round((float) (pausedTime_End[index] - pausedTime_Start[index]) / 1000);
+            final long pauseDuration = Math.round((pausedTime_End[index] - pausedTime_Start[index]) / 1000f);
             pauseDurationText = UI.format_hh_mm_ss(pauseDuration);
 
             for (; serieIndex < timeSerie.length; ++serieIndex) {
